server/requirements.txt
-i https://pypi.org/simple
bcrypt==4.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
flask==3.1.2; python_version >= '3.9'
flask-cors==6.0.1; python_version >= '3.9' and python_version < '4.0'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
setuptools==80.9.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
